expose getExistingEntities method to modules
